requirements.txt
torch
torchvision
Pillow
gdown
numpy
scipy
cmake
opencv-python
onnxruntime-gpu
